+++ Examples/RAWGKitDemo/RAWGKitDemo/RAWGKitDemoApp.swift
@@ -1,0 +1,15 @@
+//
+// RAWGKitDemoApp.swift
+// RAWGKitDemo
+//
+
+import SwiftUI
+
+@main
+struct RAWGKitDemoApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}